Se dockeriza el proyecto y se agrega api-gateway, actualizando los microservicios

+++ Payments/src/main/java/com/onieto/payments/dto/EnrollmentDto.java
@@ -2,7 +2,7 @@ package com.onieto.payments.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
+
 
 @Data
 public class EnrollmentDto {
@@ -10,7 +10,7 @@ public class EnrollmentDto {
     private Long userId;
     private Long courseId;
     private String cuponCode;
-    private BigDecimal FinalPrice;
+    private Integer FinalPrice;
 
 
 }
